+++ sites/partners/src/listings/PaperListingDetails/sections/helpers.ts
@@ -1,5 +1,5 @@
 import { t } from "@bloom-housing/ui-components"
-import moment from "moment"
+import dayjs from "dayjs"
 
 export const getDetailFieldNumber = (listingNumber: number) => {
   return listingNumber ? listingNumber.toString() : t("t.none")
@@ -10,11 +10,11 @@ export const getDetailFieldString = (listingString: string) => {
 }
 
 export const getDetailFieldDate = (listingDate: Date) => {
-  return listingDate ? moment(new Date(listingDate)).format("MM/DD/YYYY") : t("t.none")
+  return listingDate ? dayjs(new Date(listingDate)).utc().format("MM/DD/YYYY") : t("t.none")
 }
 
 export const getDetailFieldTime = (listingTime: Date) => {
-  return listingTime ? moment(new Date(listingTime)).format("hh:mm:ss A") : t("t.none")
+  return listingTime ? dayjs(new Date(listingTime)).format("hh:mm:ss A") : t("t.none")
 }
 
 export const getDetailBoolean = (listingBool: boolean) => {