+++ ui-components/src/page_components/listing/ListingsGroup.tsx
@@ -1,43 +1,41 @@
 import React, { useState } from "react"
 import { Button } from "../../actions/Button"
-import { Icon } from "../../icons/Icon"
+import { Icon, IconTypes } from "../../icons/Icon"
 import "./ListingsGroup.scss"
 
 export interface ListingsGroupProps {
   children?: React.ReactNode
-  listingsCount: number
   header: string
-  info?: string
-  showButtonText: string
   hideButtonText: string
+  icon?: IconTypes
+  info?: string
+  listingsCount: number
+  showButtonText: string
 }
 
 const ListingsGroup = (props: ListingsGroupProps) => {
   const [showListings, setShowListings] = useState(false)
   const toggleListings = () => setShowListings(!showListings)
 
-  let buttonText
-
   const listingsCount = ` (${props.listingsCount})`
-  if (showListings) {
-    buttonText = props.hideButtonText + listingsCount
-  } else {
-    buttonText = props.showButtonText + listingsCount
-  }
 
   return (
     <div className="listings-group">
       <div className="listings-group__header">
-        <div className="listings-group__icon">
-          <Icon size="xlarge" symbol="clock" />
-        </div>
-        <div className="listings-group__header-group">
-          <h2 className="listings-group__title">{props.header}</h2>
-          {props.info && <div className="px-4 my-2">{props.info}</div>}
+        <div className={"listings-group__content"}>
+          <div className="listings-group__icon">
+            <Icon size="xlarge" symbol={props.icon ?? `clock`} />
+          </div>
+          <div className="listings-group__header-group">
+            <h2 className="listings-group__title">{props.header}</h2>
+            {props.info && <div className="listings-group__info">{props.info}</div>}
+          </div>
         </div>
         <div className="listings-group__button">
           <Button className="w-full" onClick={() => toggleListings()}>
-            {buttonText}
+            {showListings
+              ? props.hideButtonText + listingsCount
+              : props.showButtonText + listingsCount}
           </Button>
         </div>
       </div>