+++ ui-components/src/page_components/listing/listing_sidebar/events/DownloadLotteryResults.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { ListingEvent } from "@bloom-housing/backend-core/types"
 import { t } from "../../../../helpers/translator"
-import moment from "moment"
+import dayjs from "dayjs"
 
 const DownloadLotteryResults = (props: { event: ListingEvent; pdfUrl: string }) => {
   const { event, pdfUrl } = props
@@ -12,7 +12,7 @@ const DownloadLotteryResults = (props: { event: ListingEvent; pdfUrl: string }) 
         <section className="aside-block text-center">
           <h2 className="text-caps pb-4">{t("listings.lotteryResults.header")}</h2>
           <p className="uppercase text-gray-800 text-tiny font-semibold pb-4">
-            {moment(event.startTime).format("MMMM D, YYYY")}
+            {dayjs(event.startTime).format("MMMM D, YYYY")}
           </p>
           <a
             className="button is-primary w-full mb-2"